requirements.txt
Flask
gunicorn
requests
Pillow
torch
accelerate
sentencepiece
protobuf
transformers
diffusers